pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "declarative-beam-pipeline"
version = "0.1.0"
description = "A declarative framework for Apache Beam pipelines using YAML configuration"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "apache-beam[gcp]>=2.40.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0",
    "importlib-metadata>=4.0.0",
    "pytest>=8.3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "pylint>=2.0.0",
]
kubernetes = [
    "kubernetes>=24.0.0",
    "jinja2>=3.0.0",
]
visualization = [
    "graphviz>=0.20.0",
]

[project.scripts]
beam-pipeline = "declarative_beam.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["declarative_beam"]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pylint.messages_control]
disable = ["C0111", "C0103", "C0330", "C0326"]
